Allow dialect methods to show extras for defs
Rather than requiring a literal string or number, the extra indicators
in the Sniff dialect will now permit references to variables bound by
defs with appropriate values as well.

A new function findDefiniteValue is defined which, given a reference to
a tile, returns either the string or number literal that is that tile or
the value of the def referred to by that var tile. If neither of these
is the case it returns null.